Add minimessage dependency
