+++ WeaponMechanics/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     compileOnly("net.kyori:adventure-api:4.11.0")
     compileOnly("net.kyori:adventure-text-serializer-legacy:4.11.0")
     compileOnly("net.kyori:adventure-platform-bukkit:4.1.1")
+    implementation("net.kyori:adventure-text-minimessage:4.11.0")
     implementation("io.lumine:Mythic-Dist:5.0.1-SNAPSHOT")
     compileOnly(files(file("../lib/crackshot/CrackShotPlus.jar")))
     compileOnly(files(file("../lib/crackshot/CrackShot.jar")))
