+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/mechanics/defaultmechanics/SoundMechanic.java
@@ -1,19 +1,26 @@
 package me.deecaad.weaponmechanics.mechanics.defaultmechanics;
 
+import me.deecaad.core.MechanicsCore;
 import me.deecaad.core.compatibility.CompatibilityAPI;
 import me.deecaad.core.file.SerializeData;
 import me.deecaad.core.file.SerializerEnumException;
 import me.deecaad.core.file.SerializerException;
 import me.deecaad.core.utils.NumberUtil;
+import me.deecaad.core.utils.ReflectionUtil;
 import me.deecaad.weaponmechanics.WeaponMechanics;
 import me.deecaad.weaponmechanics.mechanics.CastData;
 import me.deecaad.weaponmechanics.mechanics.IMechanic;
 import me.deecaad.weaponmechanics.mechanics.Mechanics;
 import me.deecaad.weaponmechanics.weapon.firearm.FirearmSound;
 import me.deecaad.weaponmechanics.weapon.reload.ReloadSound;
+import net.kyori.adventure.audience.Audience;
+import net.kyori.adventure.key.Key;
+import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import org.bukkit.Location;
 import org.bukkit.Sound;
 import org.bukkit.SoundCategory;
+import org.bukkit.World;
+import org.bukkit.entity.LivingEntity;
 import org.bukkit.scheduler.BukkitRunnable;
 
 import javax.annotation.Nonnull;
@@ -123,6 +130,7 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
                 .addArgument(double.class, false).assertArgumentRange(0.5, 2.0)
                 .addArgument(int.class, false).assertArgumentPositive()
                 .addArgument(double.class, false).assertArgumentRange(0.0, 1.0)
+                .addArgument(net.kyori.adventure.sound.Sound.Source.class, false)
                 .assertList().assertExists().get();
 
         List<SoundMechanicData> soundList = new ArrayList<>();
@@ -135,20 +143,29 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
             float pitch = split.length > 2 ? Float.parseFloat(split[2]) : 1.0f;
             int delay = split.length > 3 ? Integer.parseInt(split[3]) : 0;
             float noise = split.length > 4 ? Float.parseFloat(split[4]) : 0.0f;
+            net.kyori.adventure.sound.Sound.Source category = split.length > 5 ? net.kyori.adventure.sound.Sound.Source.valueOf(split[5]) : net.kyori.adventure.sound.Sound.Source.PLAYER;
 
             if (delay > 0)
                 delayedCounter++;
 
+            // Adventure API can play named sounds with a bunch of options, so
+            // we prefer to play custom sounds. 1.16+, all sounds are
+            // to be custom sounds.
             String stringSound = split[0].trim();
             if (stringSound.toLowerCase().startsWith("custom:")) {
                 stringSound = stringSound.substring("custom:".length());
-                soundList.add(new CustomSound(stringSound, volume, pitch, delay, noise));
+                soundList.add(new CustomSound(stringSound, volume, pitch, delay, noise, category));
                 continue;
             }
 
             try {
                 Sound sound = Sound.valueOf(stringSound);
-                soundList.add(new BukkitSound(sound, volume, pitch, delay, noise));
+
+                if (ReflectionUtil.getMCVersion() >= 16)
+                    soundList.add(new CustomSound(sound.getKey().getKey(), volume, pitch, delay, noise, category));
+                else
+                    soundList.add(new BukkitSound(sound, volume, pitch, delay, noise, category));
+
             } catch (IllegalArgumentException e) {
                 throw new SerializerEnumException(this, Sound.class, stringSound, false, data.ofList().getLocation(i));
             }
@@ -163,12 +180,14 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
         private final float pitch;
         private final int delay;
         private final float noise;
+        private final net.kyori.adventure.sound.Sound.Source category;
 
-        public SoundMechanicData(float volume, float pitch, int delay, float noise) {
+        public SoundMechanicData(float volume, float pitch, int delay, float noise, net.kyori.adventure.sound.Sound.Source category) {
             this.volume = volume;
             this.pitch = pitch;
             this.delay = delay;
             this.noise = noise;
+            this.category = category;
         }
 
         public abstract void play(CastData castData);
@@ -183,6 +202,11 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
 
         public int getDelay() {
             return delay;
+        }
+
+        @Nonnull
+        public net.kyori.adventure.sound.Sound.Source getCategory() {
+            return category;
         }
 
         public float getRandomPitch() {
@@ -202,8 +226,8 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
 
         private final Sound sound;
 
-        public BukkitSound(Sound sound, float volume, float pitch, int delay, float noise) {
-            super(volume, pitch, delay, noise);
+        public BukkitSound(Sound sound, float volume, float pitch, int delay, float noise, net.kyori.adventure.sound.Sound.Source category) {
+            super(volume, pitch, delay, noise, category);
             this.sound = sound;
         }
 
@@ -222,18 +246,27 @@ public class SoundMechanic implements IMechanic<SoundMechanic> {
 
         private final String sound;
 
-        public CustomSound(String sound, float volume, float pitch, int delay, float noise) {
-            super(volume, pitch, delay, noise);
+        public CustomSound(String sound, float volume, float pitch, int delay, float noise, net.kyori.adventure.sound.Sound.Source category) {
+            super(volume, pitch, delay, noise, category);
             this.sound = sound;
         }
 
         @Override
         public void play(CastData castData) {
-            Location castLocation = castData.getCastLocation();
-            if (CompatibilityAPI.getVersion() >= 1.11) {
-                castLocation.getWorld().playSound(castLocation, sound, SoundCategory.PLAYERS, getVolume(), getRandomPitch());
+            World world = castData.getCastLocation().getWorld();
+            net.kyori.adventure.sound.Sound adventure = net.kyori.adventure.sound.Sound.sound(Key.key(sound), getCategory(), getVolume(), getRandomPitch());
+            Audience audience = MechanicsCore.getPlugin().adventure.world(Key.key(world.getName()));
+
+            // If there is an entity involved, we can play the sound from the
+            // entity as a moving sound. In 1.15 and 1.16, the pitch/volume will
+            // be broken, but that is a small price to pay for sounds that
+            // properly track the shooter.
+            LivingEntity caster = castData.getCaster();
+            if (caster == null) {
+                Location loc = castData.getCastLocation();
+                audience.playSound(adventure, loc.getX(), loc.getY(), loc.getZ());
             } else {
-                castLocation.getWorld().playSound(castLocation, sound, getVolume(), getRandomPitch());
+                audience.playSound(adventure, (net.kyori.adventure.sound.Sound.Emitter) caster);
             }
         }
     }
